refactor: 각 파일 domain, view, controller 패키지 하위로 분리
