+++ src/main/java/step4/domain/MoveCondition.java
@@ -1,4 +1,4 @@
-package step4;
+package step4.domain;
 
 @FunctionalInterface
 public interface MoveCondition {
